requirements.txt
streamlit
pandas
numpy
plotly
gspread
google-auth
google-auth-oauthlib
cachetools
